requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiofiles==23.2.1
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
fvcore==0.1.5.post20221221
celery==5.3.6
redis==5.0.1
tortoise-orm[asyncpg]==0.20.0
aerich==0.7.2
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
